extracted source to a separate module
