+++ executor-service/build.gradle.kts
@@ -1,11 +1,10 @@
 dependencies {
-    implementation("com.squareup.okhttp3:okhttp:4.11.0")
     implementation("org.seleniumhq.selenium:selenium-java:4.13.0")
-    implementation("com.fasterxml.jackson.core:jackson-databind:2.15.2")
     implementation(project(":logger"))
     implementation(project(":model"))
     implementation(project(":webdriver"))
     implementation(project(":collection"))
+    implementation(project(":source"))
 }
