+++ executor-service/src/main/resources/application.properties
@@ -1,6 +1,2 @@
 executorservice.common.threadsCount=1
 executorservice.common.keepAliveTime=30000
-publisher.proxy.url=https://publisher-microservice-a94942799a97.herokuapp.com/publisher/proxy/count/5
-publisher.scenario.url=https://publisher-microservice-a94942799a97.herokuapp.com/publisher/scenario/count/5
-publisher.jwt.token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e30.t-IDcSemACt8x4iTMCda8Yhe3iZaWbvV5XKSTbuAn0M
-
