+++ source/src/main/java/executor/service/source/config/SourceConfiguration.java
@@ -1,16 +1,12 @@
-package executor.service.config;
+package executor.service.source.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import okhttp3.OkHttpClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-
-
 @Configuration
-public class BeanConfiguration {
-
-
+public class SourceConfiguration {
     @Bean
     public OkHttpClient getOkHttpClient() {
         return new OkHttpClient();
@@ -20,5 +16,4 @@ public class BeanConfiguration {
     public ObjectMapper objectMapper() {
         return new ObjectMapper();
     }
-
 }
