+++ source/src/main/java/executor/service/source/exception/okhttp/UnsuccessfulResponseException.java
@@ -1,4 +1,4 @@
-package executor.service.exception.source.okhttp;
+package executor.service.source.exception.okhttp;
 
 public class UnsuccessfulResponseException extends OkhttpException{
     public UnsuccessfulResponseException(int code) {
