+++ source/src/main/java/executor/service/source/listener/Header.java
@@ -1,0 +1,11 @@
+package executor.service.source.listener;
+
+import lombok.RequiredArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@RequiredArgsConstructor
+enum Header {
+    AUTHORIZATION("Authorization");
+    private final String value;
+}
