+++ source/src/main/java/executor/service/source/listener/ScenarioSourceListener.java
@@ -1,7 +1,7 @@
 package executor.service.source.listener;
 
 import executor.service.source.okhttp.AuthorizationType;
-import executor.service.model.RemoteConnection;
+import executor.service.source.model.RemoteConnection;
 import executor.service.model.Scenario;
 import executor.service.collection.queue.scenario.ScenarioSourceQueueHandler;
 import executor.service.source.okhttp.OkhttpLoader;
@@ -9,8 +9,6 @@ import okhttp3.Request;
 import org.springframework.stereotype.Component;
 
 
-
-import static com.google.common.net.HttpHeaders.AUTHORIZATION;
 @Component
 public class ScenarioSourceListener implements SourceListener {
 
@@ -22,7 +20,7 @@ public class ScenarioSourceListener implements SourceListener {
         this.loader = loader;
         this.scenarios = scenarios;
         request = new Request.Builder().url(remoteConnection.getScenarioUrl())
-                .delete().addHeader(AUTHORIZATION, AuthorizationType.BEARER.getPrefix() + remoteConnection.getToken()).build();
+                .delete().addHeader(Header.AUTHORIZATION.getValue(), AuthorizationType.BEARER.getPrefix() + remoteConnection.getToken()).build();
     }
 
     @Override
