+++ source/src/main/java/executor/service/source/model/RemoteConnection.java
@@ -1,17 +1,18 @@
-package executor.service.model;
+package executor.service.source.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Component;
 
-import java.util.Objects;
 
 @Component
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@PropertySource("classpath:connection.properties")
 public class RemoteConnection {
     @Value("${publisher.scenario.url}")
     private String scenarioUrl;
