+++ source/src/test/java/executor/service/source/okhttp/OkhttpLoaderImplTest.java
@@ -3,10 +3,10 @@ package executor.service.source.okhttp;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.type.CollectionType;
 import com.fasterxml.jackson.databind.type.TypeFactory;
-import executor.service.exception.source.DataParsingException;
-import executor.service.exception.source.okhttp.CallException;
-import executor.service.exception.source.okhttp.EmptyResponseBodyException;
-import executor.service.exception.source.okhttp.UnsuccessfulResponseException;
+import executor.service.source.exception.DataParsingException;
+import executor.service.source.exception.okhttp.CallException;
+import executor.service.source.exception.okhttp.EmptyResponseBodyException;
+import executor.service.source.exception.okhttp.UnsuccessfulResponseException;
 import okhttp3.*;
 
 import org.junit.jupiter.api.BeforeEach;
